some changes

+++ PVZ/Animation.cpp
@@ -53,6 +53,6 @@ void Animation::DrawAnimation(sf::RenderWindow& window, Coordinates positon)
 	sprite.setPosition(positon.GetX(), positon.GetY());
 	window.draw(sprite);
 }
-void Animation::ChangeScale(int Xfactor, int Yfactor) {
+void Animation::ChangeScale(float Xfactor, float Yfactor) {
 	sprite.setScale(Xfactor, Yfactor);
 }
